#514: Extract Otel setup into extension method
