+++ Api/Xipona.Api.WebApp/Program.cs
@@ -7,13 +7,7 @@ using Microsoft.Extensions.Configuration.Json;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
-using OpenTelemetry.Exporter;
-using OpenTelemetry.Logs;
-using OpenTelemetry.Metrics;
-using OpenTelemetry.Resources;
-using OpenTelemetry.Trace;
 using ProjectHermes.Xipona.Api.ApplicationServices;
 using ProjectHermes.Xipona.Api.Core;
 using ProjectHermes.Xipona.Api.Core.Files;
@@ -50,45 +44,7 @@ var configuration = builder.Configuration;
 //    .ReadFrom.Configuration(configuration)
 //    .CreateLogger();
 
-var defResourceBuilder = ResourceBuilder.CreateEmpty()
-    .AddService("Xipona.Api", serviceVersion: configuration["APP_VERSION"])
-    .AddAttributes(new Dictionary<string, object>()
-    {
-        ["deployment.environment"] = builder.Environment.EnvironmentName
-    });
-
-builder.Services
-    .AddOpenTelemetry()
-    .WithTracing(traceBuilder =>
-    {
-        traceBuilder.AddHttpClientInstrumentation();
-        traceBuilder.AddAspNetCoreInstrumentation();
-        traceBuilder.SetResourceBuilder(defResourceBuilder);
-        traceBuilder.AddSource("Example.Source"); // todo
-        traceBuilder.AddOtlpExporter(options =>
-        {
-            options.Endpoint = new Uri(configuration["OpenTelemetry:TracesEndpoint"]!);
-            options.Protocol = OtlpExportProtocol.HttpProtobuf;
-        });
-        traceBuilder.AddConsoleExporter();
-    });
-
-builder.Services.AddLogging(logging =>
-{
-    logging.AddOpenTelemetry(logOpt =>
-    {
-        logOpt.SetResourceBuilder(defResourceBuilder);
-        logOpt.IncludeFormattedMessage = true;
-        logOpt.IncludeScopes = true;
-        logOpt.ParseStateValues = true;
-        logOpt.AddOtlpExporter(opt =>
-        {
-            opt.Endpoint = new Uri(configuration["OpenTelemetry:LogsEndpoint"]!);
-            opt.Protocol = OtlpExportProtocol.HttpProtobuf;
-        });
-        logOpt.AddConsoleExporter();
-    });
-});
+builder.Services.AddOtel(configuration, builder.Environment.EnvironmentName);
 
 var fileLoadingService = new FileLoadingService();
 var vaultService = new VaultService(configuration, fileLoadingService);
